test(datetime-button): migrate disabled
